Added more utilities
added the unumlen utility and the get_unum utility. The
unumlen gets the length of an unsigned number in the base given
as argument to the function. The get_unum gets a number from the
va_list and typecasts it tot the correct type before typecastiing
back to unsigned long long (as that's the largest number a c program
can hold. Hence the reason for typecasting to unsigned long long)