Fixed a subtle bug in the printf implementation
The printf implementation before this commit did not correctly return
the length of the bytes written to stdout. It used the len variable
which was used to keep track of the current size of the buffer and was
often set to zero when the buffer size limit has been reached or
exceeded. This issue was addressed in this commit by creating another
variable to represent the length of the bytes written so far and
renaming the len variable to bufsize (which perfectly demonstrates what
the variable does).